feat: combine recipes for an easier read and output to file

+++ dsp/recipe.go
@@ -2,10 +2,10 @@ package dsp
 
 type Recipe struct {
 	OutputItem      ItemName             `json:"OutputItem"`
-	OutputItemCount float32              `json:"OutputItemCount"`
+	OutputItemCount float64              `json:"OutputItemCount"`
 	Facility        string               `json:"Facility"`
-	Time            float32              `json:"Time"`
-	Materials       map[ItemName]float32 `json:"Materials"`
+	Time            float64              `json:"Time"`
+	Materials       map[ItemName]float64 `json:"Materials"`
 }
 
 type ItemName string
